menu_planner/requirements.txt
streamlit>=1.36
python-dotenv>=1.0.0
openai>=1.0.0
pillow>=10.0
pandas>=2.2
numpy>=1.26
yfinance>=0.2.40
